src/smart_answer_core/requirements.txt
langchain==0.0.339
Markdown==3.5
markdownify==0.11.6
numpy==1.26.2
openai==0.28.1
pandas==2.1.3
pgvector==0.2.3
psycopg2_binary==2.9.9
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg-pool==3.1.8
python-dotenv==1.0.0
Requests==2.31.0
spacy==3.7.2
SQLAlchemy==2.0.22
sqlparse==0.4.4
together==0.2.8
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
